+++ NEWS.md
@@ -23,7 +23,7 @@ Breaking changes:
 - Argument conversion has been removed; the user should make sure Julia arguments passed to
   kernels match the OpenCL argument types (i.e., no empty types, 4-element tuples for
   a 3-element `float3` arguments).
-
+- The `to_host` function has been replaced by simply calling `Array` on the `CLArray`.
 
 New features: